feature/02-repo-pull-request

+++ app/src/main/java/com/viniciusjanner/apigithub/framework/data/remote/response/RepoListResponse.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 import java.io.Serializable
 
 data class RepoListResponse(
-    @SerializedName("total_count") val totalCount: Int?,
+    @SerializedName("total_count") val totalCount: Long?,
     @SerializedName("incomplete_results") val isIncompleteResults: Boolean?,
     @SerializedName("items") val items: List<ItemRepoResponse>?
 ) : Serializable
